requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
requests
numpy
faiss-cpu
torch
transformers

